OGL/SamplerCache: Treat lod_bias as a signed integer.

+++ Source/Plugins/Plugin_VideoOGL/Src/SamplerCache.cpp
@@ -99,7 +99,7 @@ void SamplerCache::SetParameters(GLuint sampler_id, const Params& params)
 	glSamplerParameterf(sampler_id, GL_TEXTURE_MAX_LOD, tm1.max_lod / 16.f);
 
 #ifndef USE_GLES3
-	glSamplerParameterf(sampler_id, GL_TEXTURE_LOD_BIAS, tm0.lod_bias / 32.f);
+	glSamplerParameterf(sampler_id, GL_TEXTURE_LOD_BIAS, (s32)tm0.lod_bias / 32.f);
 
 	if (g_ActiveConfig.iMaxAnisotropy > 0)
 		glSamplerParameterf(sampler_id, GL_TEXTURE_MAX_ANISOTROPY_EXT, (float)(1 << g_ActiveConfig.iMaxAnisotropy));
